+++ source/features/blocks/types.ts
@@ -1,9 +1,20 @@
-export interface IBlockOverview {
+export interface IBlockInfo {
+  id: string;
   number: number;
+  size: number;
+}
+
+export interface IBlockOverview extends IBlockInfo {
   createdAt: number;
   createdBy: string;
   epoch: number;
   output: number;
-  size: number;
   transactions: number;
+}
+
+export interface IBlockDetailed extends IBlockOverview {
+  confirmations: number;
+  merkleRoot: string;
+  nextBlock: string;
+  prevBlock: number;
 }