+++ source/features/epochs/types.ts
@@ -1,11 +1,17 @@
-export interface IEpochDetails {
+import { IBlockInfo } from '../blocks/types';
+
+export interface IEpochOverview {
   blocksCount: number;
   number: number;
   endedAt: Date;
-  output: number;
+  output: string;
   percentage?: number;
   slotsCount: number;
   startedAt: Date;
   status: string;
-  transactionsCount: number;
+  transactionsCount: string;
+}
+
+export interface IEpochDetails extends IEpochOverview {
+  blocks: IBlockInfo[];
 }