+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 
 ### Features
 
+- Included the mainnet theme variables and made it the default network ([PR 53](https://github.com/input-output-hk/cardano-explorer-app/pull/53))
 - Implemented table component ([PR 43](https://github.com/input-output-hk/cardano-explorer-app/pull/43))
 - Implemented stake distribution table on the current epoch page ([PR 37](https://github.com/input-output-hk/cardano-explorer-app/pull/37))
 - Implemented block creation table to for previous epoch page ([PR 36](https://github.com/input-output-hk/cardano-explorer-app/pull/36))
@@ -31,4 +32,5 @@ Changelog
 
 ### Chores
 
+- Added missing search component states ([PR 52](https://github.com/input-output-hk/cardano-explorer-app/pull/52))
 - Organized the CSS structure, global rules, and global variables ([PR 14](https://github.com/input-output-hk/cardano-explorer-app/pull/14), [PR 18](https://github.com/input-output-hk/cardano-explorer-app/pull/18))